# resources/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Golf Web Specialist 1 API",
        "version": "1.0.0",
        "description": "Development-only OpenAPI documentation for auth, profile, access, security, and notification APIs."
    },
    "servers": [
        {
            "url": "http:\/\/10.16.209.234:8000\/api\/v1",
            "description": "API v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum token"
            }
        },
...
    },
    "paths": {
...
        "\/profile\/security\/password": {
            "put": {
                "tags": [
                    "Security"
                ],
                "summary": "Update the authenticated user password.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "current_password": "password",
                                "password": "new-password",
                                "password_confirmation": "new-password"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/profile\/security\/two-factor": {
            "get": {
                "tags": [
                    "Security"
                ],
                "summary": "Get two-factor authentication status.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Enable two-factor authentication.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "force": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Security"
                ],
                "summary": "Disable two-factor authentication.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "current_password": "password"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/profile\/security\/two-factor\/confirm": {
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Confirm two-factor authentication with an authenticator code.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "code": "123456"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/profile\/security\/two-factor\/qr-code": {
            "get": {
                "tags": [
                    "Security"
                ],
                "summary": "Get the two-factor QR code SVG and otpauth URL.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
        "\/profile\/security\/two-factor\/recovery-codes": {
            "get": {
                "tags": [
                    "Security"
                ],
                "summary": "Get two-factor recovery codes.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Security"
                ],
                "summary": "Regenerate two-factor recovery codes.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application\/json": {
                                "example": {
                                    "message": "Success.",
                                    "data": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "lang",
                    "in": "query",
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum": [
                            "en",
                            "zh",
                            "es",
                            "id",
                            "de",
                            "fr"
                        ]
                    },
                    "description": "Enabled language locale used for localized response messages. For write requests, lang may also be sent in the JSON or form body when the query parameter is absent."
                }
            ]
        },
...
    }
}